clean format

+++ onnxruntime/python/tools/transformers/gpt2_beamsearch_tester.py
@@ -20,6 +20,7 @@ from benchmark_helper import Precision
 
 logger = logging.getLogger(__name__)
 
+
 class Gpt2TesterFactory:
     @staticmethod
     def create_tester(tester_type="default"):
@@ -31,37 +32,37 @@ class Gpt2TesterFactory:
         w = testers[tester_type]
         return w
 
+
 class Gpt2BeamSearchTester(Gpt2Tester):
-    def __init__(self,
-                 input_ids,
-                 position_ids,
-                 attention_mask,
-                 beam_select_idx,
-                 input_log_probs,
-                 input_unfinished_sents,
-                 prev_step_results,
-                 prev_step_scores,
-                 num_attention_heads,
-                 hidden_size,
-                 num_layer,
-                 beam_size,
-                 device,
-                 is_fp16=False,
-                 top_k=20,
-                 top_k_required_order=False,
+    def __init__(
+        self,
+        input_ids,
+        position_ids,
+        attention_mask,
+        beam_select_idx,
+        input_log_probs,
+        input_unfinished_sents,
+        prev_step_results,
+        prev_step_scores,
+        num_attention_heads,
+        hidden_size,
+        num_layer,
+        beam_size,
+        device,
+        is_fp16=False,
+        top_k=20,
+        top_k_required_order=False,
     ):
-        super().__init__(
-            input_ids,
-            position_ids,
-            attention_mask,
-            num_attention_heads=num_attention_heads,
-            hidden_size=hidden_size,
-            num_layer=num_layer,
-            device=device,
-            is_fp16=is_fp16,
-            top_k=top_k,
-            top_k_required_order=top_k_required_order
-        )
+        super().__init__(input_ids,
+                         position_ids,
+                         attention_mask,
+                         num_attention_heads=num_attention_heads,
+                         hidden_size=hidden_size,
+                         num_layer=num_layer,
+                         device=device,
+                         is_fp16=is_fp16,
+                         top_k=top_k,
+                         top_k_required_order=top_k_required_order)
         self.input_length = input_ids.shape[-1]
         self.n_layer = num_layer
         self.beam_size = beam_size
@@ -94,64 +95,41 @@ class Gpt2BeamSearchTester(Gpt2Tester):
         """
         Update the inputs for next inference.
         """
-        self.last_state = (
-            torch.from_numpy(output[0]).to(device)
-            if isinstance(output[0], numpy.ndarray)
-            else output[0].clone().detach().cpu()
-        )
+        self.last_state = (torch.from_numpy(output[0]).to(device)
+                           if isinstance(output[0], numpy.ndarray) else output[0].clone().detach().cpu())
 
         self.input_ids = self.last_state.view(self.batch_size * self.beam_size, -1).to(device)
 
         if self.position_ids is not None:
             input_unfinished_sents_id = -3
-            self.prev_step_results = (
-                torch.from_numpy(output[-2]).to(device)
-                if isinstance(output[-2], numpy.ndarray)
-                else output[-2].clone().detach().to(device)
-            )
-            self.position_ids = (
-                torch.tensor([self.input_length + step - 1])
-                .unsqueeze(0)
-                .repeat(self.batch_size * self.beam_size, 1)
-                .to(device)
-            )
+            self.prev_step_results = (torch.from_numpy(output[-2]).to(device) if isinstance(output[-2], numpy.ndarray)
+                                      else output[-2].clone().detach().to(device))
+            self.position_ids = (torch.tensor([self.input_length + step - 1
+                                               ]).unsqueeze(0).repeat(self.batch_size * self.beam_size, 1).to(device))
 
             if self.attention_mask.size(0) != (self.batch_size * self.beam_size):
-                self.attention_mask = self.attention_mask.repeat(
-                    self.batch_size * self.beam_size, 1
-                )
+                self.attention_mask = self.attention_mask.repeat(self.batch_size * self.beam_size, 1)
             self.attention_mask = torch.cat(
                 [
                     self.attention_mask,
-                    torch.ones([self.batch_size * self.beam_size, 1]).type_as(
-                        self.attention_mask
-                    ),
+                    torch.ones([self.batch_size * self.beam_size, 1]).type_as(self.attention_mask),
                 ],
                 1,
             ).to(device)
         else:
             input_unfinished_sents_id = -2
 
-        self.beam_select_idx = (
-            torch.from_numpy(output[input_unfinished_sents_id - 2]).to(device)
-            if isinstance(output[input_unfinished_sents_id - 2], numpy.ndarray)
-            else output[input_unfinished_sents_id - 2].clone().detach().to(device)
-        )
-        self.input_log_probs = (
-            torch.from_numpy(output[input_unfinished_sents_id - 1]).to(device)
-            if isinstance(output[input_unfinished_sents_id - 1], numpy.ndarray)
-            else output[input_unfinished_sents_id - 1].clone().detach().to(device)
-        )
-        self.input_unfinished_sents = (
-            torch.from_numpy(output[input_unfinished_sents_id]).to(device)
-            if isinstance(output[input_unfinished_sents_id], numpy.ndarray)
-            else output[input_unfinished_sents_id].clone().detach().to(device)
-        )
-        self.prev_step_scores = (
-            torch.from_numpy(output[-1]).to(device)
-            if isinstance(output[-1], numpy.ndarray)
-            else output[-1].clone().detach().to(device)
-        )
+        self.beam_select_idx = (torch.from_numpy(output[input_unfinished_sents_id - 2]).to(device) if isinstance(
+            output[input_unfinished_sents_id - 2], numpy.ndarray) else output[input_unfinished_sents_id -
+                                                                              2].clone().detach().to(device))
+        self.input_log_probs = (torch.from_numpy(output[input_unfinished_sents_id - 1]).to(device) if isinstance(
+            output[input_unfinished_sents_id - 1], numpy.ndarray) else output[input_unfinished_sents_id -
+                                                                              1].clone().detach().to(device))
+        self.input_unfinished_sents = (torch.from_numpy(output[input_unfinished_sents_id]).to(device) if isinstance(
+            output[input_unfinished_sents_id], numpy.ndarray) else
+                                       output[input_unfinished_sents_id].clone().detach().to(device))
+        self.prev_step_scores = (torch.from_numpy(output[-1]).to(device)
+                                 if isinstance(output[-1], numpy.ndarray) else output[-1].clone().detach().to(device))
         self.top_1_tokens = self.input_ids[0]
         self.top_k_tokens = self.last_state
 
@@ -161,11 +139,8 @@ class Gpt2BeamSearchTester(Gpt2Tester):
             self.past = list(output[1])
         else:
             for i in range(self.n_layer):
-                past_i = (
-                    torch.from_numpy(output[i + 1])
-                    if isinstance(output[i + 1], numpy.ndarray)
-                    else output[i + 1].clone().detach()
-                )
+                past_i = (torch.from_numpy(output[i + 1])
+                          if isinstance(output[i + 1], numpy.ndarray) else output[i + 1].clone().detach())
                 self.past.append(past_i.to(device))
 
     @staticmethod
@@ -222,9 +197,7 @@ class Gpt2BeamSearchTester(Gpt2Tester):
         treatment_name = "Quantized Onnx" if precision == Precision.INT8 else "Onnx"
         torch_metric = Gpt2Metric(baseline_name, baseline_name, top_k)
         onnx_metric = Gpt2Metric(treatment_name, baseline_name, top_k)
-        onnx_io_metric = Gpt2Metric(
-            treatment_name + " with IO Binding", baseline_name, top_k
-        )
+        onnx_io_metric = Gpt2Metric(treatment_name + " with IO Binding", baseline_name, top_k)
 
         for i, inputs in enumerate(test_inputs):
             if max_inputs > 0 and i == max_inputs:
@@ -233,15 +206,9 @@ class Gpt2BeamSearchTester(Gpt2Tester):
                 print(f"{i}")
             input_ids = inputs["input_ids"]
             position_ids = inputs["position_ids"] if "position_ids" in inputs else None
-            attention_mask = (
-                inputs["attention_mask"] if "attention_mask" in inputs else None
-            )
-            beam_select_idx = (
-                inputs["beam_select_idx"] if "beam_select_idx" in inputs else None
-            )
-            input_log_probs = (
-                inputs["input_log_probs"] if "input_log_probs" in inputs else None
-            )
+            attention_mask = (inputs["attention_mask"] if "attention_mask" in inputs else None)
+            beam_select_idx = (inputs["beam_select_idx"] if "beam_select_idx" in inputs else None)
+            input_log_probs = (inputs["input_log_probs"] if "input_log_probs" in inputs else None)
             input_unfinished_sents = inputs["input_unfinished_sents"]
             if model_class == "GPT2LMHeadModel_BeamSearchStep":
                 prev_step_results = inputs["input_ids"]
@@ -325,20 +292,14 @@ class Gpt2BeamSearchTester(Gpt2Tester):
                         seq_len = list(onnx_runner.input_ids.size())[-1] - past_seq_len
 
                     start_time = timeit.default_timer()
-                    pytorch_output = Gpt2BeamSearchHelper.pytorch_inference(
-                        model, torch_runner.get_inputs()
-                    )
-                    torch_metric.add_latency(
-                        past_seq_len, timeit.default_timer() - start_time
-                    )
+                    pytorch_output = Gpt2BeamSearchHelper.pytorch_inference(model, torch_runner.get_inputs())
+                    torch_metric.add_latency(past_seq_len, timeit.default_timer() - start_time)
                     torch_runner.update(pytorch_output, step, device)
 
                     (
                         onnx_output,
                         avg_latency_ms,
-                    ) = Gpt2BeamSearchHelper.onnxruntime_inference(
-                        session, onnx_runner.get_inputs(), total_runs=1
-                    )
+                    ) = Gpt2BeamSearchHelper.onnxruntime_inference(session, onnx_runner.get_inputs(), total_runs=1)
                     onnx_metric.add_latency(past_seq_len, avg_latency_ms / 1000.0)
                     onnx_runner.update(onnx_output, step, device)
 
@@ -346,7 +307,7 @@ class Gpt2BeamSearchTester(Gpt2Tester):
                         num_seq = beam_size
                     else:
                         num_seq = sum(onnx_io_runner.input_unfinished_sents.view(-1).long().cpu())
-            
+
                     output_shapes = Gpt2BeamSearchHelper.get_output_shapes(
                         batch_size,
                         context_len,
@@ -359,9 +320,7 @@ class Gpt2BeamSearchTester(Gpt2Tester):
                         num_seq=num_seq,
                     )
 
-                    Gpt2BeamSearchHelper.auto_increase_buffer_size(
-                        output_buffers, output_shapes
-                    )
+                    Gpt2BeamSearchHelper.auto_increase_buffer_size(output_buffers, output_shapes)
 
                     (
                         onnx_io_output,
@@ -375,21 +334,17 @@ class Gpt2BeamSearchTester(Gpt2Tester):
                         return_numpy=False,
                         include_copy_output_latency=True,
                     )
-                  
+
                     onnx_io_metric.add_latency(past_seq_len, avg_latency_ms / 1000.0)
 
                     if test_data_saved < save_test_data:
-                        onnx_io_runner.save_test_data(
-                            session, onnx_io_output, save_test_data_dir, test_data_saved
-                        )
+                        onnx_io_runner.save_test_data(session, onnx_io_output, save_test_data_dir, test_data_saved)
                         test_data_saved += 1
 
                     onnx_io_runner.update(onnx_io_output, step, device)
 
-                    if (
-                        (not onnx_runner.input_unfinished_sents.any()) 
-                        or (not torch_runner.input_unfinished_sents.any())
-                    ):
+                    if ((not onnx_runner.input_unfinished_sents.any())
+                            or (not torch_runner.input_unfinished_sents.any())):
                         print("break at step: ", step)
                         break
 
@@ -450,7 +405,7 @@ class Gpt2BeamSearchTester(Gpt2Tester):
                     # remove EOS
                     for k, t in enumerate(seq):
                         if t == eos_token_id:
-                            seq = seq[: k + 1]
+                            seq = seq[:k + 1]
                             break
                     print("-" * 40)
                     result = ",".join([str(token_id) for token_id in sample])
